# Base64 faster decode