--- a/Sources/FoundationEssentials/Data/Data+Base64.swift
+++ b/Sources/FoundationEssentials/Data/Data+Base64.swift
@@ -25,11 +25,6 @@ import WinSDK
 import WASILibc
 #endif
 
-private enum Base64Error: Error {
-    case invalidElementCount
-    case cannotDecode
-}
-
 @available(macOS 10.10, iOS 8.0, watchOS 2.0, tvOS 9.0, *)
 extension Data {