This PR is here to show further plans and to give a first glimpse into perf numbers that we will see. 
It should land after #1157 has landed, which will make the diff significantly smaller.