@fabianfett 2025-02-05
- base64-decode-1MB-fromData-noOptions
  - Baseline (in Foundation today): 2123μs
  - Swift implementation: 2239μs (1.05x slower)
  - This PR: 406μs (5.2x faster)
- base64-decode-1MB-fromData-noOptions-invalidAfter257bytes
  - Baseline (in Foundation today): 748ns
  - Swift implementation: 4277ns (5.7x slower)
  - This PR: 2675ns (3.5x slower)
- base64-decode-1MB-fromString-lineLength64
  - Baseline (in Foundation today): 3545μs
  - Swift implementation: 2539μs (1.4x faster)
  - This PR: 474μs (7.5x faster)
- base64-decode-1MB-fromString-noOptions
  - Baseline (in Foundation today): 3038μs
  - Swift implementation: 2232μs (1.3x faster)
  - This PR: 405μs (7.5x faster)
- base64-decode-jwtHeader-fromString-noOptions
  - Baseline (in Foundation today): 459ns
  - Swift implementation: 271ns (1.7x faster)
  - This PR: 164ns (2.8x faster)

We should investigate where we need so much time in the unhappy path. Overall nice wins!